refactor: Input 관련 내용을 ViewController로부터 InputController로 분리"
